need to cast otherwise 32bit build will fail, doh.

+++ Pulp/Common/Util/UserLiterals.h
@@ -5,5 +5,5 @@
 
 inline size_t operator "" _sz(unsigned long long int x)
 {
-	return x;
+	return static_cast<size_t>(x);
 }
